replacement of the muon propagation code
The muons in AOD lack of hit information. Hence, the code for the muon
propagation is not working any longer. It has been replaced by an official
package for the propagation of tracks to the muon station. The currently used
config files have been updated to provide all necessary information for the
propagation algorithm.
note: there is no distinction between central and forward region anymore, i.e.
the information in KDataMuon can be reduced by two floats during the next major
dictionary change. For now, both sets are filled with the same values.