avoid duplicates in tau discriminator list
Currently, the tau discriminator list, which is used for mapping tau
discriminators to bit positions, consists only of the module names. This list
neglects the process name of the discriminators. The possible duplication of
entries (if several processes have created tau discriminators) is avoided
beginning with this patch.
This modification should have no negative impact since each type of tau
discriminator should appear only once for a tau. It is notable that this
limitation was already implicitely present before this patch.