automatic trigger object selection for traditional trigger object producer
The traditional trigger object producer facilitates the saveTagsModules()
function of the HLTConfigProvider class to find all modules which produce HLT
objects. This is done for every selected trigger (as defined in the metadata
producer). Objects defined in the hard-coded list of the trigger object
producer configuration are used anyway. Be aware that this automatic lookup
only works for data that has been originally recorded in 4.2.x (or later) and
for simulated events from Fall11 (or later). The amount of data not providing
this information is small (~200/pb).