removing KMuon::sumPtIso03 and KMuon::sumPtIso05
Following the naming convention the pT sum of all tracks around a muon has
been stored in 'sumPtIso', while 'trackIso' was 0. This patch removes
'sumPtIso' and stores the same information in 'trackIso'. As a consequence
this commit not only breaks the dictionary but also the meaning of the
'trackIso' variables.