migration of tau discriminator list into separate metadata product
The tau producers are now separated from the metadata producer.
Furthermore, the tau discriminator handling has been moved to the
KTauProducer class and the template structure of the tau producer
classes has been simplified. The list of tau discriminators is stored in
a separate metadata product which requires FileInterface2 in KappaTools
for reading this information.