Rename data format classes and producers (does not compile intentionally)
As for the LVs the prefix Data is removed from the class names.
This was handled very inconsistently in the past.
As PF is default, the prefix PF is removed, too.
Now the class names are simpler and consistent.

This commit does not compile as it contains only the name changes and
no reordering and no classes.UP registration.