Use well-defined input files for automatic testing
Input files in the lower part of the HiggsTauTau configs must not use
xrootd as that is not working everywhere properly. And if changed,
the path of the output file must be changed accordingly.

Of course, you can use any of the commented out input files or others
in your private working directory.